requirements.txt
Flask
opencv-python
numpy
Flask==2.3.2
python-multipart==0.0.5
pypdf==4.0.0
pdf2image==1.17.0
Pillow==10.4.0
img2pdf==0.4.4